# visol/design/prototype/API.yaml
openapi: 3.0.0
info:
  title: ViSOL
  description: >-
    Visual Interactive Optimization Tool.
  contact: { }
  version: 1.0.0
servers:
  - url: 'https://visol1.paas.hosted-by-previder.com/api/v1'
    variables: { }
paths:
...
  /vessels:
    post:
      tags:
        - Vessels
      summary: Create a Vessel
      operationId: createVessel
      parameters:
        - in: header
          name: Reason
          schema:
            $ref: '#/components/schemas/Reason'
          required: false
          description: The reason for which the Vessel was created.
          example: New vessel inbound (#142)
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Vessel'
        required: true
      responses:
        '201':
          description: ''
          headers:
            Location:
              schema:
                type: string
                pattern: '^\/vessels\/[0-9]{1,10}$'
                maxLength: 19
                example: /vessels/3
              description: The URL of the newly created Vessel.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vessel'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /berths:
    post:
      tags:
        - Berths
      summary: Create a Berth
      operationId: createBerth
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Berth'
        required: true
      responses:
        '201':
          description: ''
          headers:
            Location:
              schema:
                type: string
                pattern: '^\/berths\/[0-9]{1,10}$'
                maxLength: 18
                example: /berths/3
              description: The URL of the newly created Berth.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Berth'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  '/terminals/{terminal_id}':
    parameters:
      - name: terminal_id
        in: path
        description: ID of the Terminal.
        required: true
        schema:
          $ref: '#/components/schemas/TerminalID'
    get:
      tags:
        - Terminals
      summary: Retrieve a Terminal
      operationId: retrieveTerminal
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Terminal'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    put:
      tags:
        - Terminals
      summary: Replace a Terminal
      operationId: replaceTerminal
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Terminal'
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Terminal'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    delete:
      tags:
        - Terminals
      summary: Delete a Terminal
      operationId: deleteTerminal
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  '/terminals/{terminal_id}/berths':
    parameters:
      - name: terminal_id
        in: path
        description: ID of the Terminal.
        required: true
        schema:
          $ref: '#/components/schemas/TerminalID'
    get:
      tags:
        - Terminals
      summary: Retrieve a Terminal's Berths
      operationId: retrieveTerminalBerths
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BerthMap'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  '/terminals/{terminal_id}/changes':
    parameters:
      - name: terminal_id
        in: path
        description: ID of the Terminal.
        required: true
        schema:
          $ref: '#/components/schemas/TerminalID'
    post:
      tags:
        - Terminals
      summary: Retrieve all changes of which the vessel's destination is the Terminal
      operationId: retrieveTerminalChanges
      parameters:
        - name: from
          in: query
          description: Time to start looking from
          schema:
            $ref: '#/components/schemas/Timestamp'
        - name: to
          in: query
          description: Time to end looking at
          schema:
            $ref: '#/components/schemas/Timestamp'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Changes'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  '/terminals/{terminal_id}/vessels':
    parameters:
      - name: terminal_id
        in: path
        description: ID of the Terminal.
        required: true
        schema:
          $ref: '#/components/schemas/TerminalID'
    get:
      tags:
        - Terminals
      summary: Retrieve all Vessels whose destination is the Terminal
      operationId: retrieveTerminalVessels
      parameters:
        - name: deadline_after
          in: query
          description: Only get Vessels who have a deadline after this time
          schema:
            $ref: '#/components/schemas/Timestamp'
        - name: arrival_before
          in: query
          description: Only get Vessels who have an arrival time before this time
          schema:
            $ref: '#/components/schemas/Timestamp'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VesselMap'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  '/terminals/{terminal_id}/schedules':
    parameters:
      - name: terminal_id
        in: path
        description: ID of the Terminal.
        required: true
        schema:
          $ref: '#/components/schemas/TerminalID'
    get:
      tags:
        - Terminals
      summary: Retrieve a Terminal's Schedules
      operationId: retrieveTerminalSchedules
      parameters:
        - name: from
          in: query
          description: Time to start looking from
          schema:
            $ref: '#/components/schemas/Timestamp'
        - name: to
          in: query
          description: Time to end looking at
          schema:
            $ref: '#/components/schemas/Timestamp'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TerminalSchedules'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  '/terminals/{terminal_id}/schedules/optimise':
    parameters:
      - name: terminal_id
        in: path
        description: ID of the Terminal.
        required: true
        schema:
          $ref: '#/components/schemas/TerminalID'
    post:
      tags:
        - Terminals
      summary: Optimise a Terminal's Schedules
      operationId: optimiseTerminalSchedules
      parameters:
        - name: from
          in: query
          description: Time to start calculating from
          schema:
            $ref: '#/components/schemas/Timestamp'
        - name: to
          in: query
          description: Time to end calculating at
          schema:
            $ref: '#/components/schemas/Timestamp'
        - name: unscheduled
          in: query
          description: Whether to include unscheduled Vessels in the optimisation
          schema:
            type: boolean
            default: false
            example: true
        - name: manual
          in: query
          description: Whether to include manual Vessels in the optimisation
          schema:
            type: boolean
            default: false
            example: true
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  '/terminals/{terminal_id}/performance':
    parameters:
      - name: terminal_id
        in: path
        description: ID of the Terminal.
        required: true
        schema:
          $ref: '#/components/schemas/TerminalID'
    get:
      tags:
        - Terminals
      summary: Calculate a Terminal's Performance
      operationId: calculateTerminalPerformance
      parameters:
        - name: from
          in: query
          description: Time to start calculating from
          schema:
            $ref: '#/components/schemas/Timestamp'
        - name: to
          in: query
          description: Time to end calculating at
          schema:
            $ref: '#/components/schemas/Timestamp'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Performance'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /changes/vessels:
    get:
      tags:
        - Changes
      summary: Retrieve all Vessels' Changes
      operationId: retrieveAllVesselChanges
      parameters:
        - name: from
          in: query
          description: Time to start looking from
          schema:
            $ref: '#/components/schemas/Timestamp'
        - name: to
          in: query
          description: Time to end looking at
          schema:
            $ref: '#/components/schemas/Timestamp'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VesselChangesMap'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /changes/schedules:
    get:
      tags:
        - Changes
      summary: Retrieve all Schedules' Changes
      operationId: retrieveAllSchedulesChanges
      parameters:
        - name: from
          in: query
          description: Time to start looking from
          schema:
            $ref: '#/components/schemas/Timestamp'
        - name: to
          in: query
          description: Time to end looking at
          schema:
            $ref: '#/components/schemas/Timestamp'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleChangesMap'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
security:
  - bearerAuth: [ ]
components:
  schemas:
    Reason:
      type: string
      pattern: '^(?:.|\s)*$'
      maxLength: 1073741824
      nullable: true
      example: New vessel inbound (#142)
    Time:
      type: string
      format: time
      pattern: '^([0-1]?[0-9]|2[0-3]):?[0-5][0-9]$'
      example: 09:30:00
    NullableTimestamp:
      allOf:
        - $ref: '#/components/schemas/Timestamp'
        - title: Nullable Timestamp
          nullable: true
    Timestamp:
      type: string
      format: date-time
      pattern: >-
        ^(-?[0-9]+)-?(1[0-2]|0[1-9])-?(3[01]|0[1-9]|[12][0-9])T(2[0-3]|[01][0-9]):?([0-5][0-9]):?([0-5][0-9])Z$
      maxLength: 22
      example: 2022-05-11T16:32:35Z
...
    TerminalSchedules:
      title: Terminal Schedules
      description: 'A map of Berth Schedules, keyed by Berth ID.'
      type: object
      additionalProperties:
        $ref: '#/components/schemas/BerthSchedules'
      example:
        '30':
          - vessel: 1
            manual: true
            berth: 30
            start: '2022-05-11T16:32:00Z'
            end: '2022-05-11T21:26:00Z'
          - vessel: 2
            manual: false
            berth: 30
            start: '2022-05-14T09:21:00Z'
            end: '2022-05-15T13:53:00Z'
...
    Performance:
      title: Performance
      description: >-
        A Performance object has the following attributes:


        + cost: `5152.25` (decimal) - The total cost of all scheduled vessels.
        
        + scheduled_vessels: `27` (integer) - The amount of vessels with a schedule.
        
        + unscheduled_vessels: `14` (integer) - The amount of vessels without a schedule.
      type: object
      properties:
        cost:
          type: number
          format: real
          minimum: 0
          maximum: 1e+37
        scheduled_vessels:
          type: integer
          format: int32
          minimum: 0
          maximum: 2147483647
        unscheduled_vessels:
          type: integer
          format: int32
          minimum: 0
          maximum: 2147483647
      additionalProperties: false
      required:
        - cost
        - scheduled_vessels
        - unscheduled_vessels
      example:
        cost: 5152.25
        scheduled_vessels: 27
        unscheduled_vessels: 14
...
    Vessel:
      title: Vessel
      description: >-
        A Vessel object has the following attributes:


        + name: `BigBoatHan` (string) - The name of the Vessel.

        + arrival: `2022-05-11T13:10:00Z` (timestamp) - When the Vessel is expected to arrive.

        + deadline: `2022-05-12T09:54:00Z` (timestamp) - When the Vessel must be completely unloaded.

        + containers: `5` (integer) - The amount of containers on the Vessel to be unloaded.

        + cost_per_hour: `2.6` (decimal) - The cost per hour from arrival till fully unloaded.

        + destination: `3` (integer) - The terminal where the Vessel should be unloaded.

        + length: `30` (integer) - The length of the Vessel in metres.

        + width: `10` (integer) - The width of the Vessel in metres.

        + depth: `20` (integer) - The depth of the Vessel in metres.
      type: object
      properties:
        name:
          type: string
          pattern: '^(?:.|\s)+$'
          maxLength: 64
        arrival:
          $ref: '#/components/schemas/Timestamp'
        deadline:
          $ref: '#/components/schemas/NullableTimestamp'
        containers:
          type: integer
          format: int32
          minimum: 0
          maximum: 2147483647
        cost_per_hour:
          type: number
          format: real
          minimum: 0
          maximum: 1e+37
        destination:
          $ref: '#/components/schemas/TerminalID'
        length:
          type: integer
          format: int32
          minimum: 0
          maximum: 2147483647
        width:
          type: integer
          format: int32
          minimum: 0
          maximum: 2147483647
        depth:
          type: integer
          format: int32
          minimum: 0
          maximum: 2147483647
      additionalProperties: false
      required:
        - name
        - arrival
        - containers
        - cost_per_hour
        - destination
        - length
        - width
        - depth
      example:
        name: Vessel 1
        arrival: 2022-05-11T13:56:00Z
        deadline: 2022-05-12T18:13:00Z
        containers: 30
        cost_per_hour: 2.6
        destination: 4
        length: 30
        width: 10
        depth: 30
    VesselMap:
      title: Vessel Map
      description: A map of Vessel objects keyed by Vessel ID.
      type: object
      additionalProperties:
        $ref: '#/components/schemas/Vessel'
      example:
        '1':
          name: Vessel 1
          arrival: 2022-05-11T13:56:00Z
          deadline: 2022-05-12T18:13:00Z
          containers: 30
          cost_per_hour: 2.6
          destination: 4
          length: 30
          width: 10
          depth: 30
        '2':
          name: The other vessel
          arrival: 2022-05-24T16:37:00Z
          deadline: 2022-05-25T23:28:00Z
          containers: 40
          cost_per_hour: 5.2
          destination: 3
          length: 45
          width: 15
          depth: 20
...
    Berth:
      title: Berth
      description: >-
        A Berth object has the following attributes:


        + terminal: `3` (integer) - The Terminal where the Berth is located.

        + open_time: `06:45` (time) - The time at which the Berth starts processing.
        
        + close_time: `21:30` (time) - The time at which the Berth stops processing.
        
        + unload_speed: `6.3` (decimal) - The amount of containers per time unit the Berth can process.
        
        + length: `150` (integer) - The length of the Berth in metres.
        
        + width: `30` (integer) - The maximum length of Vessel in the Berth in metres.
        
        + depth: `50` (integer) - The depth of the Berth in metres.
      type: object
      properties:
        terminal:
          $ref: '#/components/schemas/TerminalID'
        open_time:
          $ref: '#/components/schemas/Time'
        close_time:
          $ref: '#/components/schemas/Time'
        unload_speed:
          type: number
          format: real
          minimum: 0
          maximum: 1e+37
        length:
          type: integer
          format: int32
          minimum: 0
          maximum: 2147483647
        width:
          type: integer
          format: int32
          minimum: 0
          maximum: 2147483647
        depth:
          type: integer
          format: int32
          minimum: 0
          maximum: 2147483647
      additionalProperties: false
      required:
        - terminal
        - open_time
        - close_time
        - unload_speed
        - length
        - width
        - depth
      example:
        terminal: 3
        open_time: '06:45'
        close_time: 1290
        unload_speed: 6.3
        length: 150
        width: 30
        depth: 50
    BerthMap:
      title: Berth Map
      description: A map of Berth objects keyed by Berth ID.
      type: object
      additionalProperties:
        $ref: '#/components/schemas/Berth'
      example:
        '1':
          terminal: 3
          open_time: '06:45'
          close_time: 1290
          unload_speed: 6.3
          length: 150
          width: 30
          depth: 50
        '2':
          terminal: 3
          open_time: '08:30'
          close_time: 1155
          unload_speed: 2.7
          length: 60
          width: 20
          depth: 30
    TerminalID:
      title: Terminal ID
      description: >-
        An ID of a Terminal. Terminal IDs are unique, and are generated by the
        database.
      type: integer
      format: int32
      minimum: 1
      maximum: 2147483647
      example: 1
    Terminal:
      title: Terminal
      description: |-
        A Terminal object has the following attributes:
        
        
        + port: `2` (integer) - The Port where the Terminal is located.
        
        + name: `Han & Co. Ltd.` (string) - The name of the Terminal.
      type: object
      properties:
        port:
          $ref: '#/components/schemas/PortID'
        name:
          type: string
          pattern: '^(?:.|\s)+$'
          maxLength: 64
      additionalProperties: false
      required:
        - port
        - name
      example:
        port: 2
        name: Han & Co. Ltd.
...
    Changes:
      title: Changes
      description: An array of Vessel Change and Schedule Change objects.
      type: array
      items:
        allOf:
          - oneOf:
            - $ref: '#/components/schemas/VesselChange'
            - $ref: '#/components/schemas/ScheduleChange'
          - type: object
            properties:
             type:
              type: string
            required:
              - type
      maxItems: 2147483647
      example:
        - vessel: 30
          date: 2022-05-11T16:32:35Z
          type: 'vessel'
          old:
            name: Vessel 1
            arrival: 2022-05-11T13:56:00Z
            deadline: 2022-05-12T18:13:00Z
            containers: 30
            cost_per_hour: 2.6
            destination: 4
            length: 30
            width: 10
            depth: 30
          new:
            name: New Vessel 1
            arrival: 2022-05-11T13:56:00Z
            deadline: 2022-05-12T18:13:00Z
            containers: 40
            cost_per_hour: 2.6
            destination: 4
            length: 30
            width: 10
            depth: 30
          reason: 'Additional cargo (#2524)'
        - vessel: 30
          date: 2022-05-11T16:32:35Z
          type: 'schedule'
          old:
            vessel: 30
            manual: true
            berth: 20
            start: 2022-05-15T16:32:00Z
            end: 2022-05-15T21:26:00Z
          new:
            vessel: 30
            manual: false
            berth: 20
            start: 2022-05-15T16:32:00Z
            end: 2022-05-15T21:26:00Z
          reason: 'Turn on automatic scheduling'
...
    VesselChangesMap:
      title: Vessel Changes Map
      description: A map of Vessel Changes keyed by Vessel ID.
      type: object
      additionalProperties:
        $ref: '#/components/schemas/VesselChanges'
      example:
        '30':
          - vessel: 30
            date: 2022-05-11T16:32:35Z
            old:
              name: Vessel 1
              arrival: 2022-05-11T13:56:00Z
              deadline: 2022-05-12T18:13:00Z
              containers: 30
              cost_per_hour: 2.6
              destination: 4
              length: 30
              width: 10
              depth: 30
            new:
              name: New Vessel 1
              arrival: 2022-05-11T13:56:00Z
              deadline: 2022-05-12T18:13:00Z
              containers: 40
              cost_per_hour: 2.6
              destination: 4
              length: 30
              width: 10
              depth: 30
            reason: 'Additional cargo (#2524)'
          - vessel: 30
            date: 2022-05-11T21:26:24Z
            old:
              name: New Vessel 1
              arrival: 2022-05-11T13:56:00Z
              deadline: 2022-05-12T18:13:00Z
              containers: 40
              cost_per_hour: 2.6
              destination: 4
              length: 30
              width: 10
              depth: 30
            new:
              name: New Vessel 1
              arrival: 2022-05-18T13:56:00Z
              deadline: 2022-05-19T18:13:00Z
              containers: 40
              cost_per_hour: 2.6
              destination: 4
              length: 30
              width: 10
              depth: 30
            reason: 'Ship delayed (#3517)'
...
  responses:
    NoContent:
      description: 204 No Content
    NotFound:
      description: 404 Not Found
    Unauthorized:
      description: 401 Unauthorized
    Forbidden:
      description: 403 Forbidden
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: You must generate a JWT through the /token/obtain endpoint and then pass it as a Bearer token in the Authorization header.
...